refactor(cli): split action workflow
